LIME-14: * move enums to separate folder
