+++ frontend/src/bundles/common/components/icons/not-found-icon/components/components.ts
@@ -1,0 +1,1 @@
+export { NotFoundIcon } from './not-found-icon.component.js';
