+++ frontend/src/bundles/common/components/icons/not-found-icon/enums/enums.ts
@@ -1,0 +1,2 @@
+export { NotFoundIconColor } from './not-found-icon-color.enum.js';
+export { NotFoundIconSize } from './not-found-icon-size.enum.js';
