+++ frontend/src/bundles/common/components/icons/not-found-icon/enums/not-found-icon-color.enum.ts
@@ -1,0 +1,6 @@
+const NotFoundIconColor = {
+    PRIMARY: 'primary',
+    SECONDARY: 'secondary',
+} as const;
+
+export { NotFoundIconColor };
