+++ frontend/src/bundles/common/components/icons/not-found-icon/enums/not-found-icon-size.enum.ts
@@ -1,0 +1,8 @@
+const NotFoundIconSize = {
+    SMALL: 'sm',
+    MEDIUM: 'md',
+    LARGE: 'lg',
+    EXTRA_LARGE: 'xlg',
+} as const;
+
+export { NotFoundIconSize };
