+++ frontend/src/bundles/common/components/icons/not-found-icon/not-found-icon.ts
@@ -1,0 +1,2 @@
+export { NotFoundIcon } from './components/components.js';
+export { NotFoundIconColor, NotFoundIconSize } from './enums/enums.js';
